Move autocomplete constant to autocomplete module
