+++ ui/src/flux/constants/index.ts
@@ -1,9 +1,8 @@
 import {ast} from 'src/flux/constants/ast'
-import * as editor from 'src/flux/constants/editor'
 import * as argTypes from 'src/flux/constants/argumentTypes'
 import * as vis from 'src/flux/constants/vis'
 import * as explorer from 'src/flux/constants/explorer'
 
 const MAX_RESPONSE_BYTES = 1e7 // 10 MB
 
-export {ast, argTypes, editor, vis, explorer, MAX_RESPONSE_BYTES}
+export {ast, argTypes, vis, explorer, MAX_RESPONSE_BYTES}
