+++ ui/src/flux/helpers/autoComplete.ts
@@ -3,6 +3,66 @@ import {IInstance} from 'react-codemirror2'
 import {Suggestion} from 'src/types/flux'
 import {Hints} from 'src/types/codemirror'
 
+export const EXCLUDED_KEYS = [
+  'ArrowRight',
+  'ArrowLeft',
+  'ArrowDown',
+  'ArrowUp',
+  'Backspace',
+  'Tab',
+  'Enter',
+  'Shift',
+  'Ctrl',
+  'Control',
+  'Alt',
+  'Pause',
+  'Capslock',
+  'Escape',
+  'Pageup',
+  'Pagedown',
+  'End',
+  'Home',
+  'Left',
+  'Up',
+  'Right',
+  'Down',
+  'Insert',
+  'Delete',
+  'Left window key',
+  'Right window key',
+  'Select',
+  'Add',
+  'Subtract',
+  'Decimal point',
+  'Divide',
+  '>',
+  '|',
+  ')',
+  'F1',
+  'F2',
+  'F3',
+  'F4',
+  'F5',
+  'F6',
+  'F7',
+  'F8',
+  'F9',
+  'F10',
+  'F11',
+  'F12',
+  'Numlock',
+  'Scrolllock',
+  'Semicolon',
+  'Equalsign',
+  'Comma',
+  'Dash',
+  'Slash',
+  'Graveaccent',
+  'Backslash',
+  'Quote',
+  'Meta',
+]
+
 export const getSuggestions = (
   editor: IInstance,
   allSuggestions: Suggestion[]
